Adding the plan.
I'm building this application for many reasons.

One, it's the first good idea I've had since learning how this stuff works.
Two, I need to learn python better. I've neglected it for years.
Three, I am interested to see the ways in which a ChatGPT 4 can help me
with the planning and consult with me on my options.
-  My current theory is that GPT will be good and very helpful with a
nice greenfield application like this, but would be harder to adapt to
my process if I'm working on an existing application. This will test
that theory, because I've already tried that. It's been rather
difficult to orient GPT's help around what already exists in the
codebase without being able to just paste megabytes of code in